revive: Make other types u32

+++ cumulus/parachains/runtimes/assets/asset-hub-westend/src/lib.rs
@@ -1162,7 +1162,7 @@ parameter_types! {
 	pub const DepositPerItem: Balance = deposit(1, 0);
 	pub const DepositPerByte: Balance = deposit(0, 1);
 	pub CodeHashLockupDepositPercent: Perbill = Perbill::from_percent(30);
-	pub const BlockHashCountRevive: u64 = 256;
+	pub const BlockHashCountRevive: u32 = 256;
 }
 
 impl pallet_revive::Config for Runtime {
